Move DeepEP collectors into WideEP namespace
