+++ tests/unit/collector/test_framework_manifest.py
@@ -4,6 +4,7 @@
 """Tests for collector framework version/image manifest."""
 
 import copy
+from pathlib import Path
 
 import pytest
 
@@ -12,6 +13,9 @@ from collector.sglang.registry import REGISTRY as SGLANG_REGISTRY
 from collector.trtllm.registry import REGISTRY as TRTLLM_REGISTRY
 from collector.wideep.sglang.registry import REGISTRY as WIDEEP_SGLANG_REGISTRY
 from collector.wideep.trtllm.registry import REGISTRY as WIDEEP_TRTLLM_REGISTRY
+
+REPO_ROOT = Path(__file__).resolve().parents[3]
+COLLECTOR_ROOT = REPO_ROOT / "collector"
 
 
 def test_manifest_exposes_current_framework_versions_and_images():
@@ -62,3 +66,13 @@ def test_wideep_registry_entries_are_separate_from_stock_backend_registries():
     assert wideep_sglang_modules["wideep_mla_generation"].startswith("collector.wideep.sglang.")
     assert wideep_sglang_modules["wideep_moe"].startswith("collector.wideep.sglang.")
     assert wideep_trtllm_modules["trtllm_moe_wideep"].startswith("collector.wideep.trtllm.")
+
+
+def test_deepep_collectors_live_under_wideep_namespace():
+    assert (COLLECTOR_ROOT / "wideep" / "sglang" / "collect_deepep_moe.py").exists()
+    assert (COLLECTOR_ROOT / "wideep" / "sglang" / "deepep" / "extract_data.py").exists()
+    assert (COLLECTOR_ROOT / "wideep" / "trtllm" / "collect_moe_compute.py").exists()
+
+    assert not (COLLECTOR_ROOT / "deep_collector").exists()
+    assert not (COLLECTOR_ROOT / "sglang" / "collect_wideep_deepep_moe.py").exists()
+    assert not (COLLECTOR_ROOT / "trtllm" / "collect_wideep_moe_compute.py").exists()
